reorganize Synth and MIDIplay with their own headers
